+++ pkg/gateway/httpHandler.go
@@ -1,15 +1,15 @@
 package gateway
 
-import 	(
-  "fmt"
-  "net/http"
-  "github.com/horvatic/freighter/pkg/proxy"
+import (
+	"fmt"
+	"github.com/horvatic/freighter/pkg/proxy"
+	"net/http"
 )
 
 func handleRequest(w http.ResponseWriter, req *http.Request) {
-  response, statusCode := route(&Request{UriPath: req.URL.Path}, &proxy.RequestProxy{})
-  w.WriteHeader(statusCode)
-  fmt.Fprintf(w, response)
+	response, statusCode := route(&Request{UriPath: req.URL.Path}, &proxy.RequestProxy{})
+	w.WriteHeader(statusCode)
+	fmt.Fprintf(w, response)
 }
 
 func Start() {
